requirements.txt
numpy
astropy
ephem
networkx
sgp4
geopy
matplotlib
statsmodels
folium==0.16.0
geopandas
ortools
cartopy
git+https://github.com/snkas/exputilpy.git@v1.6
git+https://github.com/snkas/networkload.git@v1.3
pandas
h3
PyYAML
basemap
